backend/requirements.txt
# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12

# LangChain + Ollama
langchain==0.3.7
langchain-community==0.3.7
langchain-ollama==0.2.1
ollama==0.6.1

# Vector search
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1

# Document parsing
pypdf==5.1.0
python-docx==1.1.2

# Utilities
numpy==1.26.4
httpx==0.27.2
python-dotenv==1.0.1
aiofiles==24.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
